+++ gmusicapi/test/server_tests.py
@@ -21,6 +21,7 @@ from proboscis.asserts import (
 from proboscis import test, before_class, after_class, SkipTest
 
 from gmusicapi import Webclient, Musicmanager, Mobileclient
+#from gmusicapi.protocol import mobileclient
 #from gmusicapi.protocol.metadata import md_expectations
 from gmusicapi.utils.utils import retry, id_or_nid
 import gmusicapi.test.utils as test_utils
@@ -669,7 +670,7 @@ class UpauthTests(object):
         assert_equal(indices[0], pos)
 
     @plentry_test
-    def mc_reorder_ples_forwards(self):
+    def mc_reorder_ple_forwards(self):
         playlist_len = len(self.plentry_ids)
         for from_pos, to_pos in [pair for pair in
                                  itertools.product(range(playlist_len), repeat=2)
@@ -698,7 +699,7 @@ class UpauthTests(object):
                 self._mc_assert_ple_position(e_after_new_pos, to_pos + 1)
 
     @plentry_test
-    def mc_reorder_ples_backwards(self):
+    def mc_reorder_ple_backwards(self):
         playlist_len = len(self.plentry_ids)
         for from_pos, to_pos in [pair for pair in
                                  itertools.product(range(playlist_len), repeat=2)
@@ -725,6 +726,27 @@ class UpauthTests(object):
 
             if e_after_new_pos:
                 self._mc_assert_ple_position(e_after_new_pos, to_pos + 1)
+
+    # This fails, unfortunately, which means n reorderings mean n
+    # separate calls in the general case.
+    #@plentry_test
+    #def mc_reorder_ples_forwards(self):
+    #    pl = self.mc_get_playlist_songs(self.playlist_id)
+    #    # rot2, eg 0123 -> 2301
+    #    pl.append(pl.pop(0))
+    #    pl.append(pl.pop(0))
+
+    #    mutate_call = mobileclient.BatchMutatePlaylistEntries
+    #    mutations = [
+    #        mutate_call.build_plentry_reorder(
+    #            pl[-1], pl[-2]['clientId'], None),
+    #        mutate_call.build_plentry_reorder(
+    #            pl[-2], pl[-3]['clientId'], pl[-1]['clientId'])
+    #    ]
+
+    #    self.mc._make_call(mutate_call, [mutations])
+    #    self._mc_assert_ple_position(pl[-1], len(pl) - 1)
+    #    self._mc_assert_ple_position(pl[-2], len(pl) - 2)
 
     @station_test
     @all_access